Move screening information selectors to proper directory
[#22]
