+++ spec/javascripts/selectors/screening/screeningInformationFormSelectorsSpec.js
@@ -1,7 +1,7 @@
 import {
   getScreeningWithEditsSelector,
   getVisibleErrorsSelector,
-} from 'selectors/screeningInformationFormSelectors'
+} from 'selectors/screening/screeningInformationFormSelectors'
 import {fromJS, List} from 'immutable'
 import * as matchers from 'jasmine-immutable-matchers'
 import moment from 'moment'
